added screenlaunch.py
This little program manages launching services within a detached screen session. If the service crashes, the session waits for keystroke before exiting, so errors remain visible. If a screen session of the desired service already exists, you must choose to not start a new instance or replace it with a new one. In the latter case, the existing process and all children are killed softly in a INT - TERM - KILL chain.